services/retrieve-service/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
qdrant-client==1.14.3
google-genai==1.25.0